Add Databento instrument_id symbology support
