+++ nautilus_core/adapters/src/databento/symbology.rs
@@ -72,3 +72,103 @@ pub fn get_nautilus_instrument_id_for_record(
 
     Ok(InstrumentId::new(symbol, venue))
 }
+
+#[must_use]
+pub fn infer_symbology_type(symbol: &str) -> String {
+    if symbol.ends_with(".FUT") || symbol.ends_with(".OPT") {
+        return "parent".to_string();
+    }
+
+    let parts: Vec<&str> = symbol.split('.').collect();
+    if parts.len() == 3 && parts[2].chars().all(|c| c.is_ascii_digit()) {
+        return "continuous".to_string();
+    }
+
+    if symbol.chars().all(|c| c.is_ascii_digit()) {
+        return "instrument_id".to_string();
+    }
+
+    "raw_symbol".to_string()
+}
+
+pub fn check_consistent_symbology(symbols: &[&str]) -> anyhow::Result<()> {
+    if symbols.is_empty() {
+        return Err(anyhow::anyhow!("Symbols was empty"));
+    };
+
+    // SAFETY: We checked len so know there must be at least one symbol
+    let first_symbol = symbols.first().unwrap();
+    let first_stype = infer_symbology_type(first_symbol);
+
+    for symbol in symbols {
+        let next_stype = infer_symbology_type(symbol);
+        if next_stype != first_stype {
+            return Err(anyhow::anyhow!(
+                "Inconsistent symbology types: '{}' for {} vs '{}' for {}",
+                first_stype,
+                first_symbol,
+                next_stype,
+                symbol
+            ));
+        }
+    }
+
+    Ok(())
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Tests
+////////////////////////////////////////////////////////////////////////////////
+#[cfg(test)]
+mod tests {
+    use rstest::*;
+
+    use super::*;
+
+    #[rstest]
+    #[case("1", "instrument_id")]
+    #[case("123456789", "instrument_id")]
+    #[case("AAPL", "raw_symbol")]
+    #[case("ESM4", "raw_symbol")]
+    #[case("BRN FMM0024!", "raw_symbol")]
+    #[case("BRN  99   5617289", "raw_symbol")]
+    #[case("SPY   240319P00511000", "raw_symbol")]
+    #[case("ES.FUT", "parent")]
+    #[case("ES.OPT", "parent")]
+    #[case("BRN.FUT", "parent")]
+    #[case("SPX.OPT", "parent")]
+    #[case("ES.c.0", "continuous")]
+    #[case("SPX.n.0", "continuous")]
+    fn test_infer_symbology_type(#[case] symbol: String, #[case] expected: String) {
+        let result = infer_symbology_type(&symbol);
+        assert_eq!(result, expected);
+    }
+
+    #[rstest]
+    fn test_check_consistent_symbology_when_empty_symbols() {
+        let symbols: Vec<&str> = vec![];
+        let result = check_consistent_symbology(&symbols);
+        assert!(result.is_err());
+        assert_eq!(result.err().unwrap().to_string(), "Symbols was empty");
+    }
+
+    #[rstest]
+    fn test_check_consistent_symbology_when_inconsistent() {
+        let symbols = vec!["ESM4", "ES.OPT"];
+        let result = check_consistent_symbology(&symbols);
+        assert!(result.is_err());
+        assert_eq!(
+            result.err().unwrap().to_string(),
+            "Inconsistent symbology types: 'raw_symbol' for ESM4 vs 'parent' for ES.OPT"
+        );
+    }
+
+    #[rstest]
+    #[case(vec!["AAPL,MSFT"])]
+    #[case(vec!["ES.OPT,ES.FUT"])]
+    #[case(vec!["ES.c.0,ES.c.1"])]
+    fn test_check_consistent_symbology_when_consistent(#[case] symbols: Vec<&str>) {
+        let result = check_consistent_symbology(&symbols);
+        assert!(result.is_ok());
+    }
+}
